services/web/requirements.txt
Flask==2.3.2
Celery==5.3.1
gunicorn==20.0.4
psycopg2-binary==2.9.6
Flask-SQLAlchemy==3.0.5
Flask-APScheduler==1.12.4
python-dotenv==1.0.0
Authlib==1.2.1
requests==2.31.0
amqp==5.1.1
googlesearch-python==1.2.3
urllib3==2.0.4
ipinfo==5.0.0